Update patch set 1
Patch Set 1:

I agree that I probably should have uploaded my patch.  I can do so now, as it won't make a difference to the generated code, except add 4 bytes of (unused) stack space to routines that use floats but no doubles.  I was trying to avoid even that level of change, which is why I never upstreamed it.

Also, to be honest there were times in the past where more work was done in the Analyze routines, but that is no longer the case.  The whole part about the analysis was that it was supposed to be open ended.

The reason why we were treating floats and doubles differently is that it takes up to 5 instructions to materialize a double and only 2 instructions for a float, so we didn't bother for floats.

I am perfectly fine with NeedsCodebasePtr as the name, but it gets a bit messy coming up with nice names for all the helper routines.  Make the renamed AnalyzeMIR virtual, and we can both be happy

Patch-set: 1