Update patch set 1
Patch Set 1:

(3 comments)

Note this is not much about changing "mov" to "ldr", it is about making sure that the initial suspend count is a prime number. This is necessary, when profiling code running on ARM64 devices with compilation enabled. For compiled ARM64 code, we still decrement register w19 every time we do a suspend check. Setting w19 to 1000 is not a good idea.

Patch-set: 1
Reviewer: Gerrit User 1037505 <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0