Update patch set 2
Patch Set 2:

(10 comments)

There are a couple of changes unrelated to inlining. I'm ok with this since I know how convenient it is! Need to mention though that I find it easier to review when the CLs are focused :). Also, the commit message becomes more meaningful as it states exactly what you did and doesn't leave things out.

Patch-set: 2