Update patch set 3
Patch Set 3:

A majority of this patch is same as the one we are using
internally on a snapshot of AOSP which is 3-4 weeks old. We 
have a Qemu and kernel which brings up the 64-bit AOSP 
system image to the home screen with 64-bit ART in 
interpreted mode. We had run the tests with in art 
repository and have similar results as compared with other 
archs. The quick mode work is in progress. 

In AOSP, we still have some patches missing to build a 
system image. All of them are uploaded but some of them will 
need to reach AOSP through external project submissions. 
Eg: external/compiler-rt. The qemu kernel patches are all
accepted but the patches to the Qemu are still being 
reviewed.

Patch-set: 3
Label: Verified=0