Update patch set 2
Patch Set 2:

(2 comments)

I suggest adding a test to codegen_tests.cc that tries to issue a "sdiv" instruction. Please note that the ARM backend should check if the implementation has "sdiv" or not (only available on newer ARMv7 -AR CPUs - A7, A15 ... but not an A8, A9 ...) in VisitDiv(). This should fail on the target if the instruction set features are not passed correctly (only one way - on platforms that do not have support for "sdiv").

As for atomic Ldrd ... I am not aware of any strait forward way of testing this, but I will think about it. There are already some gtests in InstructionSetFeatures, you might want to take a look at those.

Patch-set: 2
Label: Code-Review=0