Update patch set 2
Patch Set 2:

Another issue is that ART isn't currently verifying that monitor's are properly scoped, see art::verifier::RegisterLine::MergeRegisters - the LOG(WARNING) should really be a verifier failure. Without this verification the following:

monitor-enter v1
if (p) {
  monitor-exit v1
}
suspend-point

would result at the suspend-point of v1 appearing locked where it wouldn't be if p were true. Sadly we see obfuscation tools generating code like this hence just logging the failures currently. In ART currently this may be an issue for our lock dumping code as part of an ANR.

I'm of course excited to hear of an implementation of the Rogers' lock!

Patch-set: 2