Update patch set 3
Patch Set 3:

Calin suggested using just '\d' to do this rather than adding non-standard sugar. '\d' is a shorthand for a digit, not a number, so we would have to write '\d+' but it would work out of the box. I guess that since numeric IDs are needed so frequently, I was looking for the shortest possible way of matching them. But maybe '\d+' is a good trade-off. What do you think, Nicolas?

Patch-set: 3
Label: Verified=0