Update patch set 2
Patch Set 2:

(3 comments)

As discussed this is an OK start, but we should remove the null checks from the graph when possible, and move the implict null check responsibility to the individual instructions.

Since you're at it, could you use the compiler options for the stack overflow checks as well?

Patch-set: 2