Update patch set 2
Patch Set 2:

A comment regarding the code generated for the null-check.
On ARM (32bit and 64bit), not considering what happens after the null-check,
the explicit version is preferable. In the general fast case (the object is not
null), the branch is predicted as not taken. A 'big' core (performing
reordering) will be able to continue execution past the load, but a 'small'
core may incur a higher penalty, especially if the memory at the address is not
in cache.

Patch-set: 2
Reviewer: Gerrit User 1052304 <1052304@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0