Update patch set 1
Patch Set 1:

(1 comment)

Mathieu will know better if this is sound, it seems questionable upfront. If a card is cleared in the alloc space but it has a reference to something on the allocation stack then at the next GC, if its a sticky GC, that alloc stack entry won't have a reference scanned that is holding it live. Its required that the allocation stack objects are added to the live set during the pause or else there would be a missing root.

Patch-set: 1
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0