Update patch set 2
Patch Set 2:

Hi, please have a look at this proposal on how to run Checker in run-test and let me know what you think. I was aiming for having to change as little as possible but maybe it could be done better. The Checker tests are still invoked the usual way, only on optimizing configs it additionally verifies the CFG.

Patch-set: 2
Label: Verified=0