Update patch set 2
Patch Set 2:

I forgot the mention you need to type something before it crashes. I think it is the "space" key or "auto complete".This fails consistently in Lollipop on both the emulator and Ingenic tablet. Basically, the problem is the StackVistor code requires that register ordering on the stack matches that of the CoreSpillMask. In this case, there were a gap between the S2 and A3 registers.

Patch-set: 2