Update patch set 3
Patch Set 3:

I need confirmation that I have to do this.  In addition, I have another x86 ABI change that I would like to upload as well, which is the 'avoid splitting long parameters between registers and stack'.

I would prefer to bump the version only once, but thought it might be better to split these up for ease in reviewing.

I think I can get that one done, but without fixes for the optimizing compiler.  Is that okay?

Patch-set: 3