Update patch set 3
Patch Set 3:

(2 comments)

Looks much better than before (I didn't do it that way because I always messed up the signature of the interpreters to five arguments, see the related comment on the DCHECK).

Looping in Mingyao as he's also working on deopt (and in fact the dex_pc).

Patch-set: 3