Update patch set 2
Patch Set 2:

Isn't the naming off? This should be continue-without-oat, right? I would also feel more comfortable if this was taken care of in ClassLinker instead. I know that dalvik_system_DexFile is now and likely in the future the only user of OpenDexFilesFromOat, but the latter still seems to be the more consistent place to do it. bdc? ngeoffray?

Patch-set: 2
Label: Code-Review=0