Update patch set 12
Patch Set 12:

confirmed fugu-userdebug not booting either.

test-art-target failures such as jni_compiler_test and compiler_driver_test as well as run-tests show the error. even with this reverted, there are other x86 test-art-target errors.

one theory for why this is not seen in test-art-host is that the target uses gcc and the host clang.

Patch-set: 12