Update patch set 2
Patch Set 2:

The current code does not prove that actually. You'd have to prove that there is no block dominating it that has a side effect and that you are dominated by the entry of the loop.

Otherwise, you can't hoist out a throwing instruction. This code works because nothing throws so you can hoist things out even if they are not being used later.

I'm still a bit unclear if this works in the case where you have an invariant in an if block that is never taken but the result is being used later:

int x = 0;

for (int i = 0; i < 10; i++) {
   if (i > 10) {
      x = 5;
   }
}

return x;

This code would not work in ART Quick since we are closer to the VRs but the optimizing compiler gets rid of the VR concept so I don't know if you'd have this problem (or if it's handled and I missed it)

Patch-set: 2