Update patch set 1
Patch Set 1:

What problem are you trying to solve? Does it justify so much added complexity?

The problems I have with how runtime options work in the existing implementation are that it's difficult to pass runtime options from within C code that are not specified as strings and that are type safe. That could be cleaned up more simply by passing ParsedOptions as the main options to the Runtime, with the option to either set the parsed options directly from within C code, or to set them using helper methods that parse string arguments.

Patch-set: 1