Update patch set 1
Patch Set 1:

Here are my thoughts based on discussion with Igor:

Motivations:
 1. Type safety for runtime options.
 2. Encapsulate how to properly parse certain kinds of things.
    For example, parsing an unsigned int is more complicated than you may think. It should be done right once and reused rather than repeatedly done wrong.
    This is also important because it makes it easier to test options parsing.
 3. Properly implement and avoid the boiler plate required for each parsed option in terms of options fields, initializers, copy/move constructors, assignments, etc.
 4. Provide a nicer syntax for specifying what the options are and how they should be parsed.

The solution to (1) is to pass parsed options to the runtime, not a vector of pairs of strings and void* pointers. Igor uses a VariantMap to hold these parsed options in a type safe way.

The solution to (2) is to write functions to parse certain kinds of types, and to reuse them when parsing options. In Igor's approach, you write the function by subclassing CmdlineType and overriding the Parse method.

The solution to (3) is to use a VariantMap, which has the boiler plate code properly implemented, and can be reused for any set of typed options you would like to collect.

The solution to (4) is to provide a builder object with Define, intoKey, and other such methods to describe the options.

In my opinion, motivations (1) and (2) are the most important, (3) and (4) not so much. Most of the complexity of the proposed change is from VariantMap, the use of subclassing to define how to parse an object of a certain type, and the builder object to make it easy to define options. If, like me, you don't think motivations (3) and (4) are that important, you can have a much simpler implementation to support (1) and (2): Use something like the ParsedOptions class already in use by the Runtime instead of a VariantMap for (1), and define ordinary functions to use for (2) rather than subclassing a class with templates. I think this will make it much easier to read, understand, and maintain the options parsing code, if slightly more tedious to write in the first place.

Patch-set: 1
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0