Update patch set 6
Patch Set 6:

(2 comments)

> (2 comments)
 > 
 > Nice.
 > 
 > What does NATIVE mean in NATIVE_COVERAGE? Why not CODE_COVERAGE?

NATIVE_COVERAGE is the native version of EMMA_INSTRUMENT (which is used to control Java coverage). It's a build level flag that must be enabled for LOCAL_NATIVE_COVERAGE to have any effect. This is done so that we don't accidentally ship coverage info if someone checks in a makefile with coverage on.

Patch-set: 6