Update patch set 1
Patch Set 1:

I am not clear how to create a regression test for this with current optimizing compiler. This bug was found when I was reviewing the code for implementing something else.

SwapLocations() is used to resolve a dependency circle in parallel move resolver. So there should be one move to stack and one move from stack to create the circle. This means we should split the use of this stack location. But currently, we are not doing this.

Patch-set: 1
Reviewer: Gerrit User 1042649 <1042649@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0