Update patch set 3
Patch Set 3:

> agampe, any further comments?

Nope, I created PS3, which makes it clear that I support a direct failure over log parsing, when necessary.

I also want this functionality for case one mentioned in the commit message, as it will clarify dex2oat aborts when we run imageless due to out-of-space or whatever (instead of the crappy bootclasspath == nullptr failure).

Patch-set: 3