Update patch set 1
Patch Set 1:

> Actually, my point is we won't create a dependency circle with
 > stack location.

that's the hope, and that's why I suggested a DCHECK.

 > 
 > BTW, there is another issue in parallel move resolver. spill
 > scratch will change sp. This makes all stack location point to the
 > wrong address.

It's not a problem of the move resolver, it's a problem of the user/backend. I haven't checked the arm64 backend, but thje others do adjust the stack if needed.

 > 
 > As I am writing another parallel move resolver, I haven't tried to
 > fix that. The original idea of the parallel move resolver is using
 > a scratch register to resolve the dependency circle instead of swap
 > so that we can emit less instructions and also try to support
 > register pair for arm32.
 > 
 > Am I going the wrong direction?

I'd rather not complicate the move resolver too much and/or have another one. If you have a simple optimization for the current one, you are going in the right direction :)

Patch-set: 1
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0