Update patch set 4
Patch Set 4:

(3 comments)

I like the idea of this cleanup but it must be implemented correctly. Using ScopedArenaVector<> will reduce peak memory usage and that's something we do care about.

However, the conversion of the switch table should be avoided (that would be extra work for arm and mips). To do that, we should record the bb or its id in Mir2Lir::SwitchTable and use bb->successor_blocks to find the labels. That could be done directly in Mir2Lir::InstallTables() so that we can even avoid the allocation of Mir2Lir::SwitchTable::targets (another small peak memory usage reduction).

Patch-set: 4
Label: Code-Review=0