Update patch set 7
Patch Set 7:

(3 comments)

The new test looks fragile because it depends on the specific wording of an error message.

Can you test instead by trying to load a dex file and execute a method from it that prints something to the screen. If something was printed, the runtime must have fallen back to the dex file, so the test fails. Otherwise nothing is printed, and the test passes. Would you still need a special check script in this case?

Patch-set: 7
Label: Verified=0