Update patch set 3
Patch Set 3:

> (3 comments)
 > 
 > 1. The original code releases the mutator lock when generating an
 > oat file (see base class_linker.cc#666). Why? Should I preserve
 > this behavior?
Actually, the original code just checked that the mutator lock was released so that somebody didn't accidentally try to compile a dex file while holding the mutator lock, which could starve GC. PS3 preserves this behavior.

 > 2. The old code checked for already opened oat files to avoid
 > loading the same oat file more than once. Should I preserve this
 > behavior?
PS3 preserves this behavior. At the very least, it reduces pressure on the virtual address space.

 > 3. Some things will be cleaner with regards to dex file ownership
 > after CL 121451 goes through.
PS3 has this.

 > 4. How should we handle a race where two processes try to generate
 > and load an oat file simultaneously? See base class_linker.cc#907
 > for reference.
I added a file-based 'Lock' method to the OatFileManager that the class linker now grabs before checking the status of and generating a new oat file. In particular, for a target oat file at /foo/bar@classes.dex, I create an empty file /foo/bar@classes.dex.flock to use for the lock.

Any feedback on this approach? I would appreciate extra eyes on the code that implements this too.

 > 5. For testing, I added code to reserve the image space before
 > loading the image. This is probably needed for other tests too, so
 > it should go into CommonRuntimeTests. But it's not entirely clear.
 > How do we know how much space to reserve and where? Why don't those
 > tests already reserve space for the image?
I'm going to leave this as it is for now, because it seems to work fine as is.

 > 6. The old code assumes dex files on the boot class path are always
 > up to date. Why? Is this right? Even in the presence of 32 versus
 > 64 architecture and boot class paths?
The OatFileManager logic doesn't apply to the dex locations in the boot class path. I'm still a little fuzzy about this. I at least think the existing behavior isn't harmful, so I've kept the behavior consistent with how it was before introducing the OatFileManager.

 > 7. Sometimes in the new code I duplicate work checking the status
 > of an oat file. The work duplicated is cheap, but the duplication
 > can lead to duplicated info log entries. This should probably be
 > avoided.
I removed duplicate work, and cleaned up the error reporting and logging. Thoughts on whether I'm logging at appropriate times with appropriate levels would be appreciated.

 > 8. We assume the state of the disk doesn't change between checking
 > the status of oat files on disk and loading the oat files. Is that
 > reasonable? In practice, I suspect this could go wrong and lead to
 > crashes. This is related to point (4).
With the new 'Lock' for the OatFileManager, this shouldn't be an issue.


I feel good about PS3. I expect to make some minor changes to handle invalid dex locations better, and I want to understand why aosp appears to take a few seconds longer to boot with my change, but otherwise, it's ready for review. All of the TODOs left in the code are things for future change lists.

Patch-set: 3