Update patch set 2
Patch Set 2:

Brian, it is not the same issue.

My case is simple, I have const bytecode and therefore I can update the workaround to make it deeper. However the bug Pavel filed is more difficult. In his case the const zero can be hided by several moves which finally can be used as fp and reference in GC map.

So to fix the bug you mentioned more work is required and to be honest we do not have the solution for that case now.

So I expect that fixing that bug will require more efforts so for a while for this specific test I'd like to extend the existed workaround.

I hope it helps to understand my intent. So do you still want I put a link to issue you mentioned?

Patch-set: 2