Update patch set 1
Patch Set 1:

(3 comments)

In cases where the warning causes you to change a variable to a more appropriate type, I think it has value.

I don't think the warning has value if you are just going to turn an implicit cast into an explicit cast. The explicit cast makes the code harder to read, and it can hide an error as easily as an implicit cast. Perhaps add a comment to the code every place there is an explicit cast to justify that it is correct?

Patch-set: 1
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0