Update patch set 3
Patch Set 3:

(9 comments)

PTAL. I've made it much stricter - the link from an instruction to its input's use list must now be always valid. As far as I know, there are two occasions when that might not be the case:
(1) SsiPhiElimination removing a Phi with a dead input
(2) DCE in HGraph::BuildDominator tree which removes instructions that may still have users (ends up calling RemoveAsUser twice).
In those cases the inputs are replaced with nullptr and ignored by RemoveAsUser.

Patch-set: 3