Update patch set 2
Patch Set 2:

PS2 adds #ifdef's in the compiler/assembly code and prevents regression in thin unlock performance in non-read-barrier configurations.

Or if you mean the whole change to be #ifdef'ed,

It could be done, but I'd be worried that #ifdef'ing the lock_word.h change (spec/comment/constants) would make the code much harder to read.

I also think it'd make a certain amount of sense to have 'sink time' in this sort of changes in stages to sanity-check (making sure all configurations work and are stable across lock word change, etc.) Making the whole thing guarded behind #ifdef would not allow this.

Patch-set: 2
Label: Verified=0