Update patch set 1
Patch Set 1:

(2 comments)

I know we discussed about this, but looking at the change, it does feel like ensuring some input consistency in a commutative operation will be simpler. We should:
1) Ensure a constant is always the last input.
2) If the two inputs are not constants, ensure they are sorted by instruction id.

We would then get most of what you wrote for free, right? Also, ensuring that a constant is always on the right improves the generated code, as most backends cannot cope with constants on the left.

Patch-set: 1
Label: Code-Review=0