Update patch set 4
Patch Set 4:

> > > I couldn't find where the dex files are actually being freed?
 > >
 > > The issue was never free-ing the dex files. When closing a dex
 > > file, you free those that are not still registered in the
 > > class-linker. The leak has always been the vector itself, which
 > is
 > > now a Java array.
 > 
 > So after the free the array holds dangling pointers? Greatness....

Free-ing is initiated from Java code. That code immediately afterwards nulls out the field storing the array, so it will be reclaimed. But if you want, I can add the code that overwrites with nulls explicitly.

 > 
 > >
 > > >
 > > > I also don't understand the need to transform the array into a
 > > > std::vector?
 > >
 > > It is simply easier than having to do the JNI array access in the
 > > places where it's used. It allows to keep most of the code as is,
 > > and isn't really much of an overhead.
 > 
 > I'm obviously not familiar with this code. Why do you need to go
 > through JNI? It's our code, can't you use mirror::Array methods?

Basically none of the code in runtime/native is going that way. Everything is using JNI, I guess to keep it in line with other native-methods code (and avoid any GC/moving stuff/deadlock issues).

Patch-set: 4