Update patch set 4
Patch Set 4:

It looks to me like this change is more about moving ownership of all DexFiles to the class linker. I would make that the primary comment in the commit message, and note that "this ensures that there can't be multiple dex files or dex caches for the same OatDexFile" as a side effect.

Specifically, the class linker owns a list of oat files, each of which owns its dex files. The class linker also owns dex files that don't belong to an oat file.

Nothing jumps out at me as to why this approach wouldn't work.
Note that this conflicts with my oat file manager change (https://android-review.googlesource.com/#/c/119862/).

I would like to go over the change in more detail before it is submitted.

Patch-set: 4