Update patch set 1
Patch Set 1:

Interesting.  On X86 & X86_64, swaps can be done with and exchange instruction, which is efficient. Looking at the test result strings, this could be less efficient for these machines. Is there a way to allow the code generator to decide to use swap or no-swap?

Patch-set: 1
Label: Verified=0