Update patch set 6
Abandoned

Vmarko, thanks for your comments. I actually agree with most of them and it seems there is quite a bit of refactoring work to make it possible. Due to lowered priority on our side, unfortunately I will not end up finishing this up. I believe the focus for select generalization should be in the Optimizing Compiler. Thus, I will abandon this patch.

Patch-set: 6
Status: abandoned