Update patch set 1
Patch Set 1:

Sorry, I forgot to ask a question yesterday:

Right now this is a neutral flag. However, once this is implemented, what are we gonna do for tests? Things like the JDWP tests might only work with the flag enabled. Should it be on by default for debug builds? Is it always the responsibility of the test writer to pass the right flag?

Patch-set: 1