Update patch set 2
Patch Set 2:

Nicolas, the diff between patch set 1 and patch set 2 better explains what I meant. 

If we collect the call positions as part of the SSA liveness analysis, we ensure that we operate on a monotonically increasing set of positions.

Otherwise we would need to sort the array before using it in RegisterAllocator::GetCallerCalleePreference().

Patch-set: 2
Label: Verified=0