Update patch set 1
Patch Set 1:

Yes, the JDWP tests should pass this flag then. 

I think we should also pass the --debuggable flag to dex2oat if the runtime is started with -Xrunjdwp option (and dex2oat needs to run to generate the oat file). This is the same than the --compiler-filter=interpret-only when starting with -Xint.

Of course, this only applies when running dalvikvm from the command-line. For app started from zygote, dex2oat must have been executed by the package manager before starting the app.

Patch-set: 1