Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

I have to agree with Nicolas that the builder (and pushing through from the graphs from the inliner) would be prettier and likely faster (as you don't need a conditional). I don't think that dead code elim removes all array accesses in practice, so I think it's OK to do it that early.

But as far as it goes, looks OK to me.

Patch-set: 1
Label: Code-Review=+1