Update patch set 3
Patch Set 3:

I think the main concern is performance. I'm not sure though how big is the impact. I think we could easily test it.

I like Andreas suggestion and I think it's not hard to implement but is also something we can work on in follow up CLs.

Add a TODO for that?

Patch-set: 3
Label: Verified=0