Update patch set 1
Patch Set 1:

That is true.  I can change the code so that it does not generate the CFI info unless the debugging flag is set.

However, before I do that, is that the right thing to do?  The CFI info is stored in the .eh_frame section, and I believe that the intention is that this section should be present in all executables.  (unlike the .debug_* sections)  ARM documentation says it should be generated even when not generating other debug tables and at high optimization levels.

Patch-set: 1