Update patch set 1
Patch Set 1:

I have found it mentioned in the following document (section 2.1.1):
http://infocenter.arm.com/help/topic/com.arm.doc.ihi0040b/IHI0040B_aadwarf.pdf

In terms of size, the .eh_frame section is about 10% of the executable.

I think that always having the unwind information present is reasonable, so that we generate back-traces on crashes, but I realize we might not want to pay the space cost.

(I will wait for the rest of the comments, and then fix them all in one go)

Patch-set: 1
Label: Verified=0