Update patch set 3
Patch Set 3: Code-Review+1

According to comment from Andreas, I have removed the counter for preload.
We only avoid to erase the dalvik-cache when we are relocating and we know that the data partition is preloaded.
This keep unchanged the error paths.
Since the erase performed in RelocateImage was not an error but done to make some space the impact is minimal.

Patch-set: 3
Label: Code-Review=+1