Update patch set 3
Patch Set 3:

Please do not use #ifdefs. We're trying hard to get away from that (as it easily leads to bitrot), at least in the current form. If you really want to do it that way, declare a bool constexpr that's defined depending on the value, and use the constexpr in the code.

There was nothing wrong with the .preloaded tag file, either. That way it is more like a runtime option instead of a compile-time one.

Patch-set: 3