Update patch set 1
Patch Set 1:

The context is the third patch, inlining shifts for long.

Without with patch, `subs` does not alter the flags, which means its result cannot be used for a `it` condition. I spent literally yesterday's afternoon, the whole night and this morning trying to understand why

__ subs(bla, bla, bla);
__ it(PL);
__ Lsr(bla, bla, bla, false, PL);

did not have the expected behaviour: executing the Lsr only if the result of the subs was positive.

You can check on the ARMv7 ref, `subs` is meant to alter flags. Without this patch, `subs()` actually emits a `subw` (as you can see in disas) which does not alter the flags, making it impractical for conditions and disrespecting the ref.

As for the tests, I don't know what you're talking about, but it's implicitely tested since we have tests for long shifts, and the implementation of long shifts depends on it.

Patch-set: 1
Reviewer: Gerrit User 1068522 <1068522@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0