Update patch set 2
Patch Set 2:

This impacts compile-time of optimizing by 10% on target and almost 2x on host. Optimizing is a heavy user of arenas, creating one per method being compiled and releasing its memory when done compiling it. Please find a way to not impact it. Let me know if you need help.

Patch-set: 2