Update patch set 1
Patch Set 1:

Hey Alexandre, Checker doesn't know the target mode because it is meant to work just on the HGraph. The only reason we dump the state after register allocation is because we also use the C1visualizer format for debugging, otherwise the output is completely arch-agnostic.

What sort of tests do you envision here? Wouldn't it be better to do arch-specific testing on assembly?

Patch-set: 1