Update patch set 2
Patch Set 2: Code-Review+2

Seems ok right now, and maybe in the future we should reconsider calling it "dex" if it's actually going to be any kind of container with a dex file inside of it. I suppose that's part of the 'improved documentation' effort we should ideally have.

Some of the terms right now, I just get the feeling they're a bit confusing to anyone not familiar with the tribal knowledge.

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2