Update patch set 4
Patch Set 4:

Thanks for your comments Nicolas (and Calin)!

 > So with this implementation, it seems to be hard to go back, in
 > case the dictionary is actually making things worse, right? Should
 > we add a bit in the CodeInfo to tell if we're using the dictionary
 > or not?

Yes, I'll implement the dynamic selection between the dictionary-based and the non-dictionary-based encoding in the next patch set (I wanted to address your other comments as well as Calin's in PS4 first).

 > Or did we actually never encounter such a case, and we should just
 > go unconditionally for it?

I do not know yet, buy that would be indeed interesting to know.  I'll check with oatdump as soon as it is updated.

Patch-set: 4
Label: Verified=0