Update patch set 7
Patch Set 7:

Yes, it's relevant for x86 and x86-64, I just started with arm64.

x86-64 has RIP-relative addressing and should be even simpler than arm64 and also save a lot of space.

For x86, we would need "call +0; pop <reg>" or a relative call to a special thunk with "mov <reg>, [esp]; ret" (I've seen such thunks in compiled C++ code) to get the EIP to a register. I would prefer promoting that to a caller-saved register instead of the ArtMethod* to reduce unnecessary EIP retrieval.

For arm, the approach should be similar to x86, except that the promoted register should point to the start of the dex cache arrays rather than just PC to reduce the number of insns typically used for the load from 3 to 2.

I didn't look at mips and mips64 in detail but I think it will benefit from the same approach as arm.

Patch-set: 7