Update patch set 1
Patch Set 1:

> > Where would you put the shared code between Quick and Optimizing,
 > > and in particular where would this CalculateMagicAndShift()
 > > function belong?
 > 
 > I am not sure. Maybe `CalculateMagicAndShift()` can be somewhere in
 > `compiler/utils`?

I don't think we have a good place for this kind of utility now. The stretch will be to put it in utils.cc but I'm not fond of it. I am happy is we leave this as a todo for now.

 > Also, some of the optimized paths are handled in constant
 > propagation (eg. 0/x) or in instruction simplification (eg. x/1).
 > Did you hit this cases?

This is a valid point. I though whether or not we should rely on the simplifier that some patterns will never arise in the codegen. First I thought that is no harm leaving the code there (there's no extra code generated). But now thinking about testing, that's a piece of code which will be untested. So IMO we should add DCHECKs in the codegen and not bother handling the cases.

Patch-set: 1