Update patch set 1
Patch Set 1:

I believe that multiplication needs to be done in double precision to get the right bits.  This may not be representable as Java type code.

You may also need to handle some edge cases, such as maxint div -1.  That is usually better done in the backend.

The rewriting of REM as div/mul/sub sounds good to me, but the special cases have to be accounted for.

Patch-set: 1