Update patch set 1
Patch Set 1:

> Why do we need another IR? I was thinking that:

I thought about collapsing div-rem in a single instruction which could be possible on x86. That requires two outputs (or a long etc...) and it's not supported with the current infrastructure.

 > - A 'normal' REM can simply be divided into DIV, MUL, SUB,. And if
 > there is an equivalent DIV, GVN will squash the two.
 > - For DIV or REM by a constant, similarly wee can just divide them
 > into shifts, add, subs, etc. Then the following GVN pass will
 > simply squash all of this.
 > I'll check with you that nobody is working on this if I decide to
 > start having a look.

it's not that simple because that doesn't work for negative numbers. And I'm not sure if adding a cmp/jmp will actually improve performance, but I did not run benchmarks...

Patch-set: 1