Update patch set 1
Patch Set 1:

> So with that, are we compiling the same set of methods? (when no
 > filter is provided)

Thanks, Nicolas.

tl;dr: We're not strictly 1:1, but I wouldn't worry about it any more.

The remaining difference is down to the fact that Quick treats 'in' vregs differently from the others, and rejects methods with vararg instructions whose vreg ranges span across the gap. To give an example, say you have a method with ten vregs and four arguments at the end. Having an invoke-*/range instruction with arguments v2-v8 is not allowed because the vreg types are not all the same. The logic is in mir_graph.cc:840.

We could easily implement similar logic in Optimizing, but I don't think it's worth it because all I found was one method in the latest GMS and one in FB, and that's it. There is also a comment in Quick which says that dx should never generate such instructions, so this behaviour should be extremely rare.

Patch-set: 1
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0