Update patch set 5
Patch Set 5:

I would obviously prefer "as is" :-) (with LLDB integration tests later)
Note that this patch does not really add any new code.  It just switches to using the DWARF library for debug_line, and it cleans up some of the code.
The intention was to split those refactorings from the big CFI CL.

Patch-set: 5
Label: Verified=0