Update patch set 1
Patch Set 1:

> why do we still need to keep the original method? What does it
 > actually test if we know it's constant folded.

Note that the original method (now named Main.HugeFrame) is not compiled, as it is deemed a "huge method", and thus does no go through constant folding (the new method, Main.$opt$LargeFrame, is simply a "large method").  I kept Main.HugeFrame as an example of a huge method that neither Optimizing nor Quick would compile, to ensure that we handle this case properly.  We may want to move it to another test though.

Patch-set: 1