Update patch set 2
Patch Set 2:

PS2 adds the same check to arm (because of thumb), x86-64 is handled with the x86 handler, and arm64 and mips have regular instruction layout.

I have to admit I don't understand why we look up dex PCs by the address of the instruction after the fault. I assume that was done to make it agree with suspend points? But that really messes us up: if somebody jumps to random memory, we'll have to try and read from there for arm and x86, and will segfault...

Patch-set: 2