Update patch set 2
Patch Set 2: Code-Review+2

This is a good first step, but you might still fail if the address isn't zero, but jumped into the middle of nowhere. The only way to protect against that is to check in the map to see if the pc is in a readable map. That is usually an expensive operation though.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2