Update patch set 1
Patch Set 1:

> I think that would be the proper thing.
 > 
 > I made the wrong assumption about kArm and kThumb2.  So are they
 > effectively an alias?
 > 
 > If they are an alias, why not remove one of them?
 > 
 > Does that mean we only support Thumb2 devices at the moment and
 > never codegen the non-thumb Arm instructions?
 > 
 > If that is the case will we ever want to support non-thumb Arm?  I
 > assume all new hardware will support Thumb2.
 > 
 > There are still some things in the code that confuse me:
 > CompiledCode::CodeDelta returns 1 for kThumb2 but not for kArm.
 > There is Arm32Assembler as well as Thumb2Assembler.

For the runtime, we don't make a difference between kThumb2 and kArm - you'll only ever see kArm in kRuntimeISA (as there's no real difference to the runtime for the devices we support).

For an ARM target, the Quick compiler silently changes it to kThumb2 (and that's stored in the CompilationUnit).

So, yes, arm devices must support thumb2 right now (which IMHO isn't a limitation, as all reasonable [from a performance point of view] CPUs support both AFAIK, e.g., it's mandatory for ARMv7 Cortex-A & -R and ARM11). The optimizing compiler might elect to support non-thumb2 devices, as the arm-assembler in compiler/utils is a pure arm32 one (as you noticed; but the utils/ assemblers are not used by the Quick backend at all). But I don't think that's wired up anywhere at all.

CodeDelta is basically the 0-bit one needs to set to tell an ARM CPU that it should run the code in thumb2 mode. The PC being odd or even is the distinguishing factor for the processor.

Patch-set: 1